S3T-510-Refactor: PlaceDto 의 PlaceItem 관련 static class 들을 PlaceItemDto 로 이동
